requirements.txt
asgiref==3.5.0
astroid==2.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
django-location-field==2.1.0
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
nose==1.3.7
passlib==1.7.4
pipreqs==0.4.11
platformdirs==2.5.1
psycopg2==2.9.3
pylint==2.12.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
yarg==0.1.9
python-dotenv==0.19.2
